Carry the port as uint16_t end to end
A port was an int in the option structs and parsePort(), a short in
NetworkServer's constructor and NetworkClient::connect() (with a comment
explaining why ports above 32767 had to be cast back before the resolver
saw them), and an unsigned short out of NetworkServer::port(), so every
hand-off cast. It is a uint16_t everywhere now: the transport's
signatures, DEFAULT_PORT and parsePort(), ServerOptions, the client's
command line and ServerConnection, the test client's session and options,
and the test helpers; the casts and the comment are gone. The connect
menu's ImGui InputInt keeps its int scratch and converts once.

The render server, MPI server and remote viewer still keep their own
short and pass it to the same constructors, converting implicitly; their
declarations are outside this document's scope.

PR review fix-up 04 item 6.

+++ src/apps/interactive/scivisStudioRemote/client/ServerConnection.cpp
@@ -38,9 +38,9 @@ namespace {
 // the UI thread never waits longer than this on the network.
 constexpr std::chrono::milliseconds COURTESY_SEND_TIMEOUT{200};
 
-std::string endpointText(const std::string &host, short port)
+std::string endpointText(const std::string &host, uint16_t port)
 {
-  return host + ":" + std::to_string(static_cast<unsigned short>(port));
+  return host + ":" + std::to_string(port);
 }
 
 } // namespace
@@ -121,7 +121,7 @@ const std::string &ServerConnection::host() const
   return m_host;
 }
 
-short ServerConnection::port() const
+uint16_t ServerConnection::port() const
 {
   return m_port;
 }
@@ -191,7 +191,7 @@ const SubtreePtr &ServerConnection::uiState() const
 
 // User intentions ////////////////////////////////////////////////////////////
 
-void ServerConnection::connect(const std::string &host, short port)
+void ServerConnection::connect(const std::string &host, uint16_t port)
 {
   if (m_state == ConnectionState::Connected) {
     vsr::core::logWarning(
